Hoàn thành cấu trúc Layout và CRUD sách

+++ qlsach/src/main/java/com/example/io/qlsach/model/Book.java
@@ -8,29 +8,49 @@ package com.example.io.qlsach.model;
  *
  * @author PC
  */
+
+/**
+ * @author PC
+ */
 public class Book {
-    // Các thuộc tính private
-    private int id;
+    // Đổi từ int sang Long để khớp với bài giảng của thầy
+    private Long id; 
     private String title;
     private String author;
 
     // Constructor không tham số
     public Book() {}
 
-    // Constructor có tham số
-    public Book(int id, String title, String author) {
+    // Constructor có tham số (Sửa int thành Long)
+    public Book(Long id, String title, String author) {
         this.id = id;
         this.title = title;
         this.author = author;
     }
 
-    // Các Getter và Setter
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    // Getter và Setter cho ID (Sửa kiểu trả về và tham số thành Long)
+    public Long getId() { 
+        return id; 
+    }
+    
+    public void setId(Long id) { 
+        this.id = id; 
+    }
 
-    public String getTitle() { return title; }
-    public void setTitle(String title) { this.title = title; }
+    public String getTitle() { 
+        return title; 
+    }
+    
+    public void setTitle(String title) { 
+        this.title = title; 
+    }
 
-    public String getAuthor() { return author; }
-    public void setAuthor(String author) { this.author = author; }
+    public String getAuthor() { 
+        return author; 
+    }
+    
+    public void setAuthor(String author) { 
+        this.author = author; 
+    }
 }
+
